// extensions/conditional-shipping-discount/generated/input.ts
export type DiscountClass =
  /**
   * The discount is combined with an
   * [order discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'ORDER'
  /**
   * The discount is combined with a
   * [product discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'PRODUCT'
  /**
   * The discount is combined with a
   * [shipping discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  | 'SHIPPING';

export type CartDeliveryOptionsDiscountsGenerateRunInputVariables = Exact<{ [key: string]: never; }>;


export type CartDeliveryOptionsDiscountsGenerateRunInput = { cart: { lines: Array<{ id: string }>, cost: { subtotalAmount: { amount: string, currencyCode: CurrencyCode } }, deliveryGroups: Array<{ deliveryOptions: Array<{ handle: string, title: string | null }> }> }, discount: { discountClasses: Array<DiscountClass>, config: { jsonValue: unknown } | null } };